+++ src/CviConverter/dto/Led.cs
@@ -8,9 +8,10 @@ namespace CviConverter.dto
 {
     public class Led : BasePanel
     {
+
         public string xtype { get; set; } = "fieldset";
-        public double x { get; set; }
-        public double y { get; set; }
+        public int x { get; set; }
+        public int y { get; set; }
         public string style { get; set; }
         public List<VStyle> vstyle { get; set; } = new List<VStyle>();
         public string cls { get; set; }
